requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==38.0.3
Cython==0.29.32
defusedxml==0.7.1
Django==4.1.1
django-bootstrap3==22.2
django-cors-headers==3.13.0
django-filter==22.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
google==3.0.0
httpie==3.2.1
idna==3.4
Markdown==3.4.1
markdown-it-py==2.1.0
mdurl==0.1.2
multidict==6.0.4
mysqlclient==2.1.1
Pillow==9.4.0
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
pymssql==2.2.5
PyMySQL==1.0.2
PySocks==1.7.1
pytz==2022.4
requests==2.28.2
requests-toolbelt==0.10.1
rich==13.3.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
urllib3==1.26.14
uWSGI==2.0.21
